apps/backend/requirements.txt
# Core data/ML & apps
numpy
pandas
scikit-learn
streamlit

# APIs & LLM tooling
groq
openai
langchain
chromadb

# Web framework & server
fastapi
uvicorn

# HTTP & parsing
requests
beautifulsoup4
lxml

# Env
python-dotenv==1.0.1

# Realtime / sockets
websockets

# Database ORM & driver
SQLAlchemy==2.0.31
psycopg[binary]>=3.2,<4    # PostgreSQL driver (v3) – no compiler needed on 3.13

# Auth & uploads
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
